+++ src/main/java/com/revature/RevRelay/repository/UserRepository.java
@@ -1,8 +1,6 @@
-package com.revature.RevRelay.persistence;
+package com.revature.RevRelay.repository;
 
 import com.revature.RevRelay.models.User;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.Optional;
